feat: 支持H264、H265的视频流解析自定义SEI数据

+++ src/demuxer/demuxer.cpp
@@ -545,9 +545,9 @@ bool Demuxer::handleLoopPlayback()
 
     // 重置到文件开始位置
     if (formatContext_->pb && formatContext_->pb->seekable) {
-        const int ret = avio_seek(formatContext_->pb, 0, SEEK_SET);
+        const int64_t ret = avio_seek(formatContext_->pb, 0, SEEK_SET);
         if (ret < 0) {
-            LOG_ERROR("{} Seek to start failed: {}", url_, utils::avErr2Str(ret));
+            LOG_ERROR("{} Seek to start failed: {}", url_, utils::avErr2Str(static_cast<int>(ret)));
             return false;
         }
     }
@@ -676,9 +676,9 @@ bool Demuxer::openInternal(const std::string &url, const Config &config,
 
     // 重置到文件开始位置
     if (formatContext_->pb && formatContext_->pb->seekable) {
-        const int ret = avio_seek(formatContext_->pb, 0, SEEK_SET);
+        const int64_t ret = avio_seek(formatContext_->pb, 0, SEEK_SET);
         if (ret < 0) {
-            LOG_ERROR("{} Seek to start failed: {}", url_, utils::avErr2Str(ret));
+            LOG_ERROR("{} Seek to start failed: {}", url_, utils::avErr2Str(static_cast<int>(ret)));
             return false;
         }
     }
